Protocol: add disablepoweroff at code 0x0a, and debug_out at code 0x0b

+++ src/ascii_protocol.c
@@ -72,13 +72,13 @@ extern int pwms[2];
 
 
 
-extern int debug_out;
-extern int disablepoweroff;
+extern uint8_t debug_out;
+extern uint8_t disablepoweroff;
 extern int powerofftimer;
 extern uint8_t buzzerFreq;    // global variable for the buzzer pitch. can be 1, 2, 3, 4, 5, 6, 7...
 extern uint8_t buzzerPattern; // global variable for the buzzer pattern. can be 1, 2, 3, 4, 5, 6, 7...
 extern int buzzerLen;
-extern int enablescope; // enable scope on values
+extern uint8_t enablescope; // enable scope on values
 
 static int speedB = 0;
 static int steerB = 0;
