+++ src/comms.c
@@ -24,8 +24,8 @@ extern UART_HandleTypeDef huart3;
 volatile uint8_t uart_buf[100];
 volatile int16_t ch_buf[8];
 
-int debug_out = 1;
-int enablescope = 1;
+uint8_t debug_out = 1;
+uint8_t enablescope = 1;
 
 //volatile char char_buf[300];
